Redesign per-thread schema information
Pull all RLMRealm-specific data out of RLMObjectSchema and RLMProperty (i.e.
the table on RLMObjectSchema and the column on RLMProperty) and add a new
object which stores that and pointers to both versions of the object schema.

Do all table column lookups via the realm::Property rather than caching the
column in various places to support the column numbers changing at any time
(and at *different* times on different threads).

Share RLMSchema instances between all RLMRealms for a given path, and remove
the accessor cache in favor of just storing the accessor classes on the
RLMSchema object.

Remove the logic to skip creating accessor classes for read-only Realms which
are missing tables, as that was originally just a workaround for the fact that
the accessor classes would be bogus and cause problems when read-write Realms
tried to use the same class. This no longer applies.